Rename component references in tests
